OSW-12: docs: add swagger to authentication api
