+++ src/app/api/auth/otpRegistration/route.ts
@@ -8,7 +8,7 @@ import CommonService from '@/services/commonService';
  *   post:
  *     tags:
  *       - Authentication
- *     description: Send OTP to user's email.
+ *     description: Send OTP to user's email to authenticate user's email to register.
  *     requestBody:
  *       required: true
  *       content:
@@ -19,9 +19,14 @@ import CommonService from '@/services/commonService';
  *               email:
  *                 type: string
  *                 example: bddquan@gmail.com
+ *               username:
+ *                  type: string
+ *                  example: Design_ITUS
  *     responses:
  *       200:
- *         description: Send otp successfully.The otp is returned in a cookie named `OTP` which expired in 5 minutes. You need to include this cookie in `verifyOTP` requests.
+ *         description: Send otp successfully.The otp is returned in a cookie named `OTP` which expired in 5 minutes. You need to include this cookie in `registration` requests.
+ *       409:
+ *          description: Existed email or username.
  *       500:
  *         description: Error message.
  */
